Rename `kontrol` module and command to `kontrolx`
In order to avoid name collision with the downstream project in the
interim period both projects distribute the module.
