+++ kevm-pyk/pyproject.toml
@@ -11,7 +11,7 @@ authors = [
 ]
 packages = [
     { include = "kevm_pyk", from = "src/" },
-    { include = "kontrol", from = "src/" },
+    { include = "kontrolx", from = "src/" },
 ]
 
 [tool.poetry.dependencies]
@@ -43,7 +43,7 @@ kevm = "kevm_pyk.__main__:main"
 kevm-pyk = "kevm_pyk.__main__:main"
 kevm-dist = "kevm_pyk.dist:main"
 gst-to-kore = "kevm_pyk.gst_to_kore:main"
-kontrol = "kontrol.__main__:main"
+kontrolx = "kontrolx.__main__:main"
 
 [tool.isort]
 profile = "black"
